kafka-simple-python/pyproject.toml
[tool.poetry]
name = "kafka-simple-python"
version = "0.0.1"
description = ""
authors = ["paullee714 <paul_lee@kakao.com>"]

[tool.poetry.dependencies]
python = "^3.9"
kafka-python = "^2.0.2"
black = "^21.7b0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
